refactor: move CheckProps type definition to a new iProps file for better organization
